apps/api/requirements.txt
-e ../agent
fastapi
uvicorn
httpx
pydantic
python-dotenv
langgraph
